+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Change `ContainerCreateOptsBuilder::restart_policy` function to take a strongly typed `opts::container::ContainerRestartPolicy`
 * Rename `opts::RestartPolicy` -> `opts::SystemdRestartPolicy` to avoid name collision
 * Fix `Container::attach` when connecting to a container that has allocated TTY
+* Fix error handling in `Images::build`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync